// src/lib/interfaces/componentNames.ts
export type ComponentName =
  | "BAGTAG_PRINTER"
  | "BAGTAG_FEEDER"
  | "BAGTAG_DISPENSER"
  | "BOARDINGPASS_PRINTER"
  | "BOARDINGPASS_FEEDER"
  | "BOARDINGPASS_DISPENSER"
  | "PASSPORT_READER"
  | "BARCODE_READER"
  | "GENERAL_PURPOSE_PRINTER"
  | "GENERAL_PURPOSE_FEEDER"
  | "GENERAL_PURPOSE_DISPENSER"
  | "ILIMINATION"
  | "ADA_PAD"
  | "ANNOUNCEMENT"
  | "DISPLAY"
  | "PAYMENT"
  | "CONVEYOR"
  | "BIOMETRIC";

export const ComponentName = {
  BAGTAG_PRINTER: "BAGTAG_PRINTER" as ComponentName,
  BAGTAG_FEEDER: "BAGTAG_FEEDER" as ComponentName,
  BAGTAG_DISPENSER: "BAGTAG_DISPENSER" as ComponentName,
  BOARDINGPASS_PRINTER: "BOARDINGPASS_PRINTER" as ComponentName,
  BOARDINGPASS_DISPENSER: "BOARDINGPASS_DISPENSER" as ComponentName,
  PASSPORT_READER: "PASSPORT_READER" as ComponentName,
  BARCODE_READER: "BARCODE_READER" as ComponentName,
  GENERAL_PURPOSE_PRINTER: "GENERAL_PURPOSE_PRINTER" as ComponentName,
  GENERAL_PURPOSE_FEEDER: "PASSPORT_READER" as ComponentName,
  GENERAL_PURPOSE_DISPENSER: "GENERAL_PURPOSE_DISPENSER" as ComponentName,
  ILIMINATION: "ILIMINATION" as ComponentName,
  ADA_PAD: "ADA_PAD" as ComponentName,
  ANNOUNCEMENT: "ANNOUNCEMENT" as ComponentName,
  DISPLAY: "DISPLAY" as ComponentName,
  PAYMENT: "PAYMENT" as ComponentName,
  CONVEYOR: "CONVEYOR" as ComponentName,
  BIOMETRIC: "BIOMETRIC" as ComponentName
};